backend/requirements.txt
# Core FastAPI & Web Framework (Your original versions)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
scikit-learn>=1.3.2

# --- ADDED FOR STREAMLIT UI ---
streamlit==1.28.1
plotly==5.18.0
altair==5.1.2

# --- ADDED FOR BACKEND COMPATIBILITY ---
httpx==0.25.1           # Required by FastAPI TestClient
python-dateutil==2.8.2  # Date handling for pandas
pytz==2023.3.post1      # Timezone support

# --- OPTIONAL BUT USEFUL ---
scipy==1.11.4           # Often needed by scikit-learn
matplotlib==3.8.2       # For additional plotting
seaborn==0.13.0         # Statistical visualizations